+++ src/panelWebView/components/Preview.tsx
@@ -5,20 +5,12 @@ import { ActionButton } from './ActionButton';
 import * as l10n from '@vscode/l10n';
 import { LocalizationKey } from '../../localization';
 
-export interface IPreviewProps {
-  slug: string;
-}
+export interface IPreviewProps { }
 
-const Preview: React.FunctionComponent<IPreviewProps> = ({
-  slug
-}: React.PropsWithChildren<IPreviewProps>) => {
+const Preview: React.FunctionComponent<IPreviewProps> = (_: React.PropsWithChildren<IPreviewProps>) => {
   const open = () => {
     Messenger.send(CommandToCode.openPreview);
   };
-
-  if (!slug) {
-    return null;
-  }
 
   return <ActionButton onClick={open} title={l10n.t(LocalizationKey.panelPreviewTitle)} />;
 };